Tslm 159 rewrite st bigdata consume batch ma with cr ecd to be idempotent pipeline

+++ script/main.py
@@ -1,4 +1,6 @@
+import json
 import sys
+from datetime import datetime, timedelta
 
 import boto3
 from awsglue import DynamicFrame
@@ -6,12 +8,17 @@ from awsglue.context import GlueContext
 from awsglue.job import Job
 from awsglue.transforms import *
 from awsglue.utils import getResolvedOptions
-from helper import GlobalVariables, Helper, Queries
+from dateutil.relativedelta import relativedelta
+from helper import Helper, Queries
 from pyspark.context import SparkContext
+from pyspark.sql import DataFrame
 from pyspark.sql import functions as F
 
 
 def parse_env_name() -> str:
+    """
+    Parse the environment name from the account alias
+    """
     client = boto3.client("iam")
 
     account_alliases = client.list_account_aliases()
@@ -20,17 +27,24 @@ def parse_env_name() -> str:
     return env_name
 
 
-ENV_NAME = parse_env_name()
-
-
-def cacheDf(df, glueContext, transformation_ctx) -> DynamicFrame:
+def cacheDf(
+    df: DynamicFrame, glueContext: GlueContext, transformation_ctx: str
+) -> DynamicFrame:
+    """
+    Cache the dynamicframe to optimize performance
+    """
     cached_df = df.toDF()
     cached_df = cached_df.cache()
 
     return DynamicFrame.fromDF(cached_df, glueContext, transformation_ctx)
 
 
-def sparkUnion(glueContext, unionType, mapping, transformation_ctx) -> DynamicFrame:
+def sparkUnion(
+    glueContext: GlueContext, unionType: str, mapping: dict, transformation_ctx: str
+) -> DynamicFrame:
+    """
+    Spark union operation
+    """
     for alias, frame in mapping.items():
         frame.toDF().createOrReplaceTempView(alias)
     result = spark.sql(
@@ -40,8 +54,15 @@ def sparkUnion(glueContext, unionType, mapping, transformation_ctx) -> DynamicFr
 
 
 def sparkSqlQuery(
-    glueContext, query, mapping, transformation_ctx, plan=False
+    glueContext: GlueContext,
+    query: str,
+    mapping: dict,
+    transformation_ctx: str,
+    plan: bool = False,
 ) -> DynamicFrame:
+    """
+    Spark SQL query operation
+    """
     for alias, frame in mapping.items():
         frame.toDF().createOrReplaceTempView(alias)
     result = spark.sql(query)
@@ -50,28 +71,53 @@ def sparkSqlQuery(
     return DynamicFrame.fromDF(result, glueContext, transformation_ctx)
 
 
-def update_delete(glueContext, df) -> DynamicFrame:
+def filter_red_red(df: DynamicFrame) -> DynamicFrame:
+    """
+    Filter red_red dataframe
+    """
+    df = df.toDF()
+    result = df.filter(
+        (F.col("cleaned_classified_distributionType").isin("RENT", "BUY"))
+        & (
+            F.col("classified_geo_countrySpecific_de_iwtLegacyGeoID").startswith("108")
+            | F.col("classified_geo_countrySpecific_de_iwtLegacyGeoID").startswith(
+                "103"
+            )
+        )
+        & (F.col("classified_estateType").isin("HOUSE", "APARTMENT"))
+    )
+    return DynamicFrame.fromDF(result, glueContext, "red_red_filtered")
+
+
+def update_delete(glueContext: GlueContext, df: DynamicFrame) -> DynamicFrame:
+    """
+    Clean red_red deleted records and prepare it
+    """
     columns = df.toDF().columns
-    prefixes = ["classified_", "cleaned_", "cleanup", "extracted_", "grenzwert_"]
 
     # Use dictionary comprehension to create lists of columns based on prefixes
-    filtered_columns_wo_prefix = [
-        x for prefix in prefixes for x in columns if x.startswith(prefix)
-    ]
-    filtered_columns_str_wo_prefix = ",".join(filtered_columns_wo_prefix)
-
-    filtered_columns_with_prefix = [
-        "b." + x for prefix in prefixes for x in columns if x.startswith(prefix)
-    ]
-    filtered_columns_str_with_prefix = ",".join(filtered_columns_with_prefix)
+    filtered_columns_str_wo_prefix = ",".join(
+        x
+        for prefix in config["validPrefixValues"]
+        for x in columns
+        if x.startswith(prefix)
+    )
+    filtered_columns_str_with_prefix = ",".join(
+        "b." + x
+        for prefix in config["validPrefixValues"]
+        for x in columns
+        if x.startswith(prefix)
+    )
 
     ret_df = sparkSqlQuery(
         glueContext,
         query=Queries.get_merge_delete_query(
             extra_columns_wo_prefix=filtered_columns_str_wo_prefix,
             extra_columns_with_prefix=filtered_columns_str_with_prefix,
+            first_day_past=first_day_past,
+            partition_date=partition_date,
         ),
-        mapping={"red_red_cleaned": df},
+        mapping={"red_red_filtered": df},
         transformation_ctx="merged_df",
     )
     ret_df = ret_df.drop_fields(paths=["rank"])
@@ -79,10 +125,44 @@ def update_delete(glueContext, df) -> DynamicFrame:
     return cacheDf(ret_df, glueContext, "cached_update_delete")
 
 
-def modifyData(glueContext, df, geoid) -> DynamicFrame:
+def join_csv_static_data(df: DataFrame) -> DataFrame:
+    bundeslaender_df = spark.read.csv(
+        "bundeslaender.csv", header=True, inferSchema="true"
+    )
+    stadtlandkreise_df = spark.read.csv(
+        "stadtlandkreise.csv", header=True, inferSchema="true"
+    )
+
+    ret_df = (
+        df.join(
+            F.broadcast(bundeslaender_df),
+            F.substring(df.classified_geo_countrySpecific_de_iwtLegacyGeoID, 1, 5)
+            == bundeslaender_df.geoid,
+            how="left",
+        )
+        .drop("geoid")
+        .join(
+            F.broadcast(stadtlandkreise_df),
+            F.substring(df.classified_geo_countrySpecific_de_iwtLegacyGeoID, 1, 8)
+            == stadtlandkreise_df.geoid,
+            how="left",
+        )
+        .drop("geoid", "stadtkreis")
+        .withColumnRenamed("bundesland", "geo_state")
+        .withColumnRenamed("landkreis", "geo_userDefined_immoWelt_county")
+    )
+    return ret_df
+
+
+def modify_data(
+    glueContext: GlueContext, df: DynamicFrame, geoid: int, partition_month: str
+) -> DynamicFrame:
+    """
+    Custom transformation for the data
+    """
     ret_df = df.toDF()
 
-    # clean values
+    # Clean the columns
     ret_df = ret_df.withColumn(
         "cleaned_classified_structure_rooms_numberofrooms",
         F.round(F.col("cleaned_classified_structure_rooms_numberofrooms")).cast("int"),
@@ -90,79 +170,100 @@ def modifyData(glueContext, df, geoid) -> DynamicFrame:
         "classified_geo_city",
         F.regexp_replace(F.col("classified_geo_city"), "\\\\\\\\", ""),
     )
-    # rename columns
+
+    # Remove the prefix from the column names
     cols_to_rename = [c for c in ret_df.columns if c.startswith("cleaned_")]
     for old_name in cols_to_rename:
         new_name = old_name.replace("cleaned_", "")
         ret_df = ret_df.withColumnRenamed(old_name, new_name)
 
     if geoid == 108:
-        bundeslaender_df = spark.read.csv(
-            "bundeslaender.csv", header=True, inferSchema="true"
-        )
-        stadtlandkreise_df = spark.read.csv(
-            "stadtlandkreise.csv", header=True, inferSchema="true"
-        )
-
-        ret_df = (
-            ret_df.join(
-                bundeslaender_df,
-                F.substring(
-                    ret_df.classified_geo_countrySpecific_de_iwtLegacyGeoID, 1, 5
-                )
-                == bundeslaender_df.geoid,
-                how="left",
-            )
-            .drop("geoid")
-            .join(
-                stadtlandkreise_df,
-                F.substring(
-                    ret_df.classified_geo_countrySpecific_de_iwtLegacyGeoID, 1, 8
-                )
-                == stadtlandkreise_df.geoid,
-                how="left",
-            )
-            .drop("geoid", "stadtkreis")
-            .withColumnRenamed("bundesland", "geo_state")
-            .withColumnRenamed("landkreis", "geo_userDefined_immoWelt_county")
-        )
+        ret_df = join_csv_static_data(ret_df)
     else:
         ret_df = ret_df.withColumn("geo_state", F.lit(None))
         ret_df = ret_df.withColumn("geo_userDefined_immoWelt_county", F.lit(None))
 
-    ret_df = ret_df.drop(
-        "geoid",
-        "classified_geo_userDefined_immoWelt_geoid",
-        "classified_geo_countrySpecific_de_iwtLegacyGeoID",
-    )
+    ret_df = ret_df.drop(*config["geoDropColumns"])
 
     # add additional columns
     ret_df = ret_df.withColumn("partitionGeoid", F.lit(geoid)).withColumn(
-        "partitionMonth", F.lit(GlobalVariables.partition_month)
+        "partitionMonth", F.lit(partition_month)
     )
 
     return DynamicFrame.fromDF(ret_df, glueContext, "attributes")
 
 
-args = getResolvedOptions(sys.argv, ["JOB_NAME"])
+def set_date_values(
+    partition_date: datetime.date, days_ago: str
+) -> tuple[str, str, str, str]:
+    """
+    Set date values for the queries
+    """
+    first_day_current_month = partition_date.replace(day=1).strftime("%Y-%m-%d")
+    if days_ago == "full_refresh":
+        first_day_past = "2024-05-20"  # oldest date available in the data
+    else:
+        first_day_past = (
+            partition_date.replace(day=1) + relativedelta(days=-int(days_ago))
+        ).strftime("%Y-%m-%d")
+
+    partition_month = partition_date.strftime("%Y-%m")
+
+    return (
+        first_day_current_month,
+        partition_date.strftime("%Y-%m-%d"),
+        first_day_past,
+        partition_month,
+    )
+
+
+# load arguments
+args = getResolvedOptions(sys.argv, ["JOB_NAME", "partition_date", "days_ago"])
 sc = SparkContext()
 glueContext = GlueContext(sc)
+
 spark = glueContext.spark_session
 job = Job(glueContext)
 job.init(args["JOB_NAME"], args)
 
-########## get base tables from data catalog
-red_red_cleaned_draft = glueContext.create_dynamic_frame.from_catalog(
-    database="kafka",
-    table_name="red_red_cleaned",
-    transformation_ctx="red_red_cleaned_draft",
+ENV_NAME = parse_env_name()
+
+# load config
+with open("config.json") as f:
+    config = json.load(f)
+
+if args["partition_date"] == "yesterday":
+    partition_date = (datetime.now() - timedelta(days=1)).date()
+else:
+    partition_date = datetime.strptime(args["partition_date"], "%Y-%m-%d")
+
+full_refresh = True if args.get("days_ago") == "full_refresh" else False
+# set date values
+(
+    first_day_current_month,
+    partition_date,
+    first_day_past,
+    partition_month,
+) = set_date_values(partition_date=partition_date, days_ago=args["days_ago"])
+
+# getting base tables from data catalog
+red_red_cleaned_raw = glueContext.create_dynamic_frame.from_options(
+    connection_type="s3",
+    format="parquet",
+    connection_options={
+        "paths": [f"s3://ingest-stream-red-red-{ENV_NAME}/data/cleaned_new"],
+        "recurse": True,
+    },
+    transformation_ctx="red_red_cleaned_raw",
 )
-red_red_cleaned = update_delete(glueContext=glueContext, df=red_red_cleaned_draft)
+
+red_red_filtered = filter_red_red(red_red_cleaned_raw)
+red_red_cleaned = update_delete(glueContext=glueContext, df=red_red_filtered)
 
 red_red_text = glueContext.create_dynamic_frame.from_catalog(
     database="kafka",
     table_name="red_red_text",
-    push_down_predicate=f"(partitioncreateddate>=to_date('{GlobalVariables.first_day_3_months_ago}') and partitioncreateddate<to_date('{GlobalVariables.first_day_next_month}'))",
+    push_down_predicate=f"(partitioncreateddate>=to_date('{first_day_past}') and partitioncreateddate<=to_date('{partition_date}'))",
     transformation_ctx="red_red_text",
 )
 
@@ -178,39 +279,38 @@ red_vd_cleaned = DynamicFrame.fromDF(
 red_ecd = glueContext.create_dynamic_frame.from_catalog(
     database="kafka",
     table_name="red_ecd_raw",
-    push_down_predicate=f"(partitioncreateddate>=to_date('{GlobalVariables.first_day_3_months_ago}') and partitioncreateddate<to_date('{GlobalVariables.first_day_next_month}'))",
+    push_down_predicate=f"(partitioncreateddate>=to_date('{first_day_past}') and partitioncreateddate<=to_date('{partition_date}'))",
     transformation_ctx="red_ecd",
 )
 
 contactrequests_daily_cr_per_classified = glueContext.create_dynamic_frame.from_catalog(
     database="kinesis",
     table_name="contactrequests_daily_cr_per_classified",
-    push_down_predicate=f"(partitioncreateddate>=to_date('{GlobalVariables.first_day_current_month}') and partitioncreateddate<to_date('{GlobalVariables.first_day_next_month}'))",
+    push_down_predicate=f"(partitioncreateddate>=to_date('{first_day_current_month}') and partitioncreateddate<=to_date('{partition_date}'))",
     transformation_ctx="contactrequests_daily_cr_per_classified",
 )
 
 customeractions_daily_actions_per_classified = glueContext.create_dynamic_frame.from_catalog(
     database="kinesis",
     table_name="customeractions_daily_actions_per_classified",
-    push_down_predicate=f"(partitioncreateddate>=to_date('{GlobalVariables.first_day_current_month}') and partitioncreateddate<to_date('{GlobalVariables.first_day_next_month}'))",
+    push_down_predicate=f"(partitioncreateddate>=to_date('{first_day_current_month}') and partitioncreateddate<=to_date('{partition_date}'))",
     transformation_ctx="customeractions_daily_actions_per_classified",
 )
-##########
-
-# MAIN
-
-# GEOID, DATA_COUNTRY, DISTRIBUTION_TYPE, DATA_SOURCE
-country_values = [
-    (108, "GERMANY", "BUY", "SELL"),
-    (108, "GERMANY", "RENT", "RENTAL"),
-    (103, "AUSTRIA", "BUY", "SELL"),
-    (103, "AUSTRIA", "RENT", "RENTAL"),
-]
+# end of getting base tables
 
 union_df = None
-for geoid, data_country, distribution_type, data_source in country_values:
-    print(f"Getting data for: {geoid, data_country, distribution_type, data_source}")
-    queries_obj = Queries(distribution_type=distribution_type, geoid=geoid)
+
+
+# loop over possible values
+for row in config["countryValues"]:
+    geoid = row["geoid"]
+    country_name = row["country_name"]
+    distribution_type = row["distribution_type"]
+    data_source = row["data_source"]
+
+    queries_obj = Queries(
+        distribution_type, geoid, first_day_current_month, partition_date
+    )
 
     BaseDataFirst = sparkSqlQuery(
         glueContext=glueContext,
@@ -222,7 +322,6 @@ for geoid, data_country, distribution_type, data_source in country_values:
         transformation_ctx="BaseData_first",
     )
     BaseDataFirst = cacheDf(BaseDataFirst, glueContext, "basedata_first_cached")
-    print("Done fetching base data first")
 
     BaseData = sparkSqlQuery(
         glueContext=glueContext,
@@ -236,8 +335,6 @@ for geoid, data_country, distribution_type, data_source in country_values:
         },
         transformation_ctx="BaseData_df",
     )
-    BaseData = cacheDf(BaseData, glueContext, "basedata_cached")
-    print("Done fetching base data cached")
 
     BaseData_final_df = sparkSqlQuery(
         glueContext=glueContext,
@@ -248,15 +345,16 @@ for geoid, data_country, distribution_type, data_source in country_values:
         },
         transformation_ctx="BaseData_final_df",
     )
-    print("Done fetching base data final")
 
     BaseData_final_df = BaseData_final_df.drop_fields(
-        paths=GlobalVariables.cols_to_drop_basedata
+        paths=config["colsToDropBaseData"]
     )
-    BaseData_final_df = modifyData(
+
+    BaseData_final_df = modify_data(
         glueContext=glueContext,
         df=BaseData_final_df,
         geoid=geoid,
+        partition_month=partition_month,
     )
 
     if union_df is None:
@@ -269,14 +367,14 @@ for geoid, data_country, distribution_type, data_source in country_values:
             transformation_ctx="Union_node",
         )
 
-    print("Union df done")
+    BaseDataFirst.toDF().unpersist()
 
-    csv_df = BaseData_final_df.drop_fields(paths=GlobalVariables.cols_to_drop_json)
-    json_df = Helper.modifyDataJson(
+    csv_df = BaseData_final_df.drop_fields(paths=config["colsToDropJson"])
+    json_df = Helper.modify_data_json(
         glueContext=glueContext, df=csv_df, distribution_type=distribution_type
     )
 
-    s3_path_json = f"s3://consume-batch-ma-with-cr-ecd-{ENV_NAME}/data/{data_country.lower()}/{distribution_type.lower()}/json/partitioncreateddate={GlobalVariables.today}"
+    s3_path_json = f"s3://consume-batch-ma-with-cr-ecd-{ENV_NAME}/data/{country_name.lower()}/{distribution_type.lower()}/json/partitioncreateddate={'full_refresh' if full_refresh else partition_date}"
     AmazonS3_node1714127201181 = glueContext.write_dynamic_frame.from_options(
         frame=json_df,
         connection_type="s3",
@@ -285,7 +383,7 @@ for geoid, data_country, distribution_type, data_source in country_values:
         transformation_ctx="AmazonS3_node1714127201181",
     )
 
-    s3_path_csv = f"s3://consume-batch-ma-with-cr-ecd-{ENV_NAME}/data/{data_country.lower()}/{distribution_type.lower()}/csv/partitioncreateddate={GlobalVariables.today}"
+    s3_path_csv = f"s3://consume-batch-ma-with-cr-ecd-{ENV_NAME}/data/{country_name.lower()}/{distribution_type.lower()}/csv/partitioncreateddate={'full_refresh' if full_refresh else partition_date}"
     AmazonS3_node1714127201182 = glueContext.write_dynamic_frame.from_options(
         frame=csv_df.coalesce(1),
         connection_type="s3",
@@ -293,30 +391,28 @@ for geoid, data_country, distribution_type, data_source in country_values:
         connection_options={"compression": "gzip", "path": s3_path_csv},
         transformation_ctx="AmazonS3_node1714127201181",
     )
-    print("Done with json")
 
-# delete insert instead of replacewhere
-glueContext.purge_table(
-    database="kafka",
-    table_name="offers_ma_geo",
-    options={
-        "partitionPredicate": f"(partitionmonth == '{GlobalVariables.partition_month}')",
-        "retentionPeriod": 1,
-    },
-)
-print("Done purging table")
+if not full_refresh:
+    # delete insert instead of replacewhere
+    glueContext.purge_table(
+        database="kafka",
+        table_name="offers_ma_geo",
+        options={
+            "partitionPredicate": f"(partitionmonth == '{partition_month}')",
+            "retentionPeriod": 1,
+        },
+    )
 
-AWSGlueDataCatalog_node1709799333156 = glueContext.write_dynamic_frame.from_catalog(
-    frame=union_df,
-    database="kafka",
-    table_name="offers_ma_geo",
-    additional_options={
-        "enableUpdateCatalog": True,
-        "updateBehavior": "UPDATE_IN_DATABASE",
-        "partitionKeys": ["partitionMonth"],
-    },
-    transformation_ctx="AWSGlueDataCatalog_node1709799333156",
-)
-print("Done writing to table")
+    AWSGlueDataCatalog_node1709799333156 = glueContext.write_dynamic_frame.from_catalog(
+        frame=union_df,
+        database="kafka",
+        table_name="offers_ma_geo",
+        additional_options={
+            "enableUpdateCatalog": True,
+            "updateBehavior": "UPDATE_IN_DATABASE",
+            "partitionKeys": ["partitionMonth"],
+        },
+        transformation_ctx="AWSGlueDataCatalog_node1709799333156",
+    )
 
 job.commit()
